src/main/java/br/com/dicasdeumdev/api/services/impl/UserServiceImpl.java
package br.com.dicasdeumdev.api.services.impl;

import br.com.dicasdeumdev.api.domain.User;
import br.com.dicasdeumdev.api.domain.dto.UserDTO;
import br.com.dicasdeumdev.api.repositories.UserRepository;
import br.com.dicasdeumdev.api.services.exceptions.DataIntregityViolationException;
import br.com.dicasdeumdev.api.services.exceptions.ObjectNotFoundException;
import org.modelmapper.ModelMapper;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {

    @Autowired
    private UserRepository repository;

    @Autowired
    private ModelMapper modelMapper;

    @Override
    public User findById(Long id) {
        Optional<User> obj = repository.findById(id);
        return obj.orElseThrow(() -> new ObjectNotFoundException("Usuario com o ID nao encontrado!"));
    }

    @Override
    public List<User> findAll() {
        return repository.findAll();
    }

    @Override
    public User create(UserDTO obj) {
        findByEmail(obj);
        return repository.save(modelMapper.map(obj, User.class));
    }

    @Override
    public void findByEmail(UserDTO obj) {
        Optional<User> user = Optional.ofNullable(repository.findByEmail(obj.getEmail()));
        if (user.isPresent() && !user.get().getId().equals(obj.getId())) {
            throw new DataIntregityViolationException("Email já cadastrado");
        }
    }

    @Override
    public User update(UserDTO obj) {
        findByEmail(obj);
        return repository.save(modelMapper.map(obj, User.class));
    }

    @Override
    public void delete(Long id) {
        findById(id);
        repository.deleteById(id);
    }
}